feat: unified results fetcher
